+++ src/lib/oidc/index.ts
@@ -1,4 +1,4 @@
 export { AuthSecure } from './AuthSecure';
-export { useAccessToken } from './useAccessToken ';
+export { useAccessToken } from './useAccessToken';
 export { useAuth } from './useAuth';
 export { useAuthUser } from './useAuthUser';
